Suppress debugging output
So far, the debugging output is just dumped onto cout. This change
suppresses all output so that really wanted output can be properly
tested (and not drowned under the river of debugging output).

For now, the output is commented out, but a better logging scheme
should be on the TODO.