Add an interface specification for ObjectApproximators
ObjectApproximators will be classes that provide approximations
of given point clouds. The approximators do not need to know
whether the given point cloud is a part of a larger point cloud.

The approximators should be stateless and not have any other
dependencies to facilitate their testing and future concurrency.

For now, the interface itself contains a dependency on some of
the legacy code, namely, the ObjectModel. This is in order to
make a step-by-step refactoring where the initial introduction of
the approximator keeps the same functionality as previously.