Adapt the legacy approximation to the new interface
Migrate the legacy moment-of-inertia based approximation to the
new interface.

The code for finding the coefficients of the particular models is
mostly copy-pasted from the original implementation (only the
dependencies adapted).

In the future, the models themselves should know how to find the
coefficients based on the given descriptors, so that the approximator
can only generate the descriptor, delegate to the models, and finally
choose the model with the lowest error. This, however, would be
a bigger refactoring and so is left out for now.