Implement initial splitting of objects
For now, only a single iteration is performed where each object
found in the original cloud is split once and the split approximated.

All the machinery is set for when there is a better way to
evaluate whether a cloud should be further split
(based on some error of the current approximation).