Add an interface for point cloud segmentation
This interface specifies which methods need to be implemented
by classes that implement point cloud segmentation.

Importantly, the new interface mandates that the cloud that
should be segmented is passed in as a parameter to the
segmenter, rather than having the segmenter pull it from the
parent instance (like in the legacy code).