Port the legacy segmenter to the new interface
The legacy segmentation algorithm relies on performing Euclidean
plane extraction until there are no more planes in the original
cloud, after which it performs EuclideanClustering to obtain the
segments that should be considered objects within the original
cloud.

The relevant code is simply factored into the new implementation,
rather than being a rewrite of the algorithm, to make sure that
the interface refactoring keeps the original functionality
unchanged.