Rewrite the Euclidean-plane-based segmentation algorithm
The legacy implementation was riddled with many problems and
inefficiencies.  Unused structures and code paths, very
messy code, misleading (at best) and wrong comments, etc.

This rewrite keeps the same approach to segment the original
point cloud into clouds that are considered obstacle clouds,
but does so with a much cleaner implementation and, as such,
provides not only cosmetic benefits to the code base, but
also efficiency (speed-ups) and efficacy (more precise segment
isolation) improvements.