Implement a filtered video source
The filtered video source is implemented as a `VideoSource` decorator.

This decorator is both a video observer and a video source: it observes
a given video source and provides (as a source) a cloud obtained by
filtering the received clouds.

For now, the filtering performed by this decorator is hard coded, but
it could be adapted to take a list of filter steps that should be
applied to the cloud as a parameter in the future, once it's clearer
which filters are going to be used.