Implement obstacle output smoothening
A `SmoothObstacleAggregator` decorator of an `ObstacleAggregator`
performs a simple output smoothening by propagating only objects
that have been seen enough frames in a row to the underlying
aggregator and removing tracked obstacles only if they are lost
enough frames in a row.

This way, "flickering" of detected obstacles should be avoided in
a large number of cases.

The implementation assumes that any detected obstacles are already
in the same coordinate system, regardless of the frame they were
detected in. In other words, if the camera moves between the frames,
the detected obstacles' positions need to be adjusted for the camera
movement or else the smoothener is unable to discern that two objects
are the same one (since their coordinates differ between frames).