Implement a LOLA-specific obstacle aggregator
This aggregator sends the received obstacles to the LOLA controller
wrapped in a UDP datagram.

The obstacles are serialized by representing each obstacle with 11
integers and packing their representations in series. The peer can
decode the total number of obstacles by dividing the payload size
by (11 * 4).

The integers are encoded using system-specific endianess, i.e.
not the network (big) endian. This assumption is due to the existing
LOLA system working this way and forcing a big endian representation
would introduce extra coordination problems.

For now, the LolaAggregator is implemented in a single header file,
even though it's not a templated class. This is because LOLA-specifc
code should eventually be moved out of the lepp tree (or at least
to a sibling-level where the LOLA-specific detector executable would
be implemented).