Periodically update the approximation shapes
The blending process would only make sure the position of the
approximation tracks (smoothly) the position of the detected
object. With this change, we periodically update the shapes too
(by completely replacing the old ones with the new ones).

Note: this is quite a hacky implementation, but until it proves
itself as useful, it's better than refactoring the obstacle
aggregation interface to take `CompositeModel`s instead of the
more generic `ObjectModel`s.