Factor out the splitting logic of object approximation
The splitting logic is currently mixed with the logic that generates
an approximation based on the principal component analysis. This,
however, should not be the case as these two things are completely
orthogonal: splitting the object can be done independently of how each
part's approximation is found; objects can be approximated by using the
principal component analysis without doing any splitting.