Add a SplitStrategy interface
This interface will allow us to implement different algorithms for
defining how and when point clouds are split during the
`SplitObjectApproximator`'s object approximation process, without
having to change (or recompile) the `SplitObjectApproximator` itself.

This decouples the approximation algorithm that performs splits of
the cloud that it is approximating from the algorithm that does the
splitting.