Extract the RobotService interface to an ABC
The previously existing RobotService is now renamed to
`AsyncRobotService` and is an implementation of the interface.

This change makes it possible to easily implement alternate robot
services (for sending vision messages) without having to change the
clients (that only rely on having a reference to the ABC). One of the
possibilities is now to create real unit tests that mock out the robot
service.