Do not modify or delete objects in the inner zone of the robot
Since the robot could get confused in the middle of a step if we were to
modify or delete objects that are found in its inner zone, we defer
deletions until the robot has moved sufficiently away from the object
and stop modifying it as long as the robot moves close enough.

This works since the robot's assumption is that nothing in its inner
zone can move (a nearly static environment).