LOLA: Implement a general Context container
The `Context` is a container class that owns all parts of the object
detection (and in the future the entire vision subsystem) pipeline.

The `Context` ABC provides convenience methods for setting up the
pipeline, by factoring out various parts of its configuration into small
self-contained methods.

Concrete implementations of the `Context` can be provided that configure
it in varying ways, such as a `HardcodedContext` (included in this
commit) or a `ConfigFileContext` (future work).