Implement file-based config of the vision subsystem
The `FileConfigContext` class implements the `Context` interface and
reads its configuration from a toml-based config file. The config file
is given in TOML syntax, but the parsing is not based on a fully
compliant toml parser, making the order of the config values
significant.

The vision subsystem executable can now be invoked by either passing it
a config file (in which case it is fully configured by parsing the
config file) or as previously, by passing some command line parameters
(in which case some values are left hardcoded in the executable binary).
In general, the config-file based invocations should be preferred.