Remove unnecessary and **invalid** message len assignment
GCC seems to allow this, assignment, although the function is static and
the `params` name shouldn't really be in scope (it's a member of the
class, but the function is static, so...). Clang returns an error for
this.

Additionally, it's unnecessary to even assign the length there, as the
default ctor already makes sure that the length is correctly set.