Added Circular Buffer to Avoid Missing Any Frame
Due to the nature of cloud_cb_ method in PCL::Grabber interface, all
incoming frames (clouds) from the OpenNIGrabber interface are passed
to our callback in one thread. This might cause the cloud_cb_ method
to lose some of these clouds without knowing it. Since the
frame_counter_  are loosely connected in the current implementation,
FileOdoTransfomer and GrabberVideoSource classes can easily lose
track of the current incoming cloud.
To Tackle this issue, a circular buffer is introduced and used as a
temporary container. It keeps all the incoming frames and a separate
thread will be in charge of processing each cloud in that buffer.
This will avoid losing any clouds and ensuring the consistency of the
frame numbers passed between different classes.