Stairviz Mimics Same Behavior as ObstacleViz
For the sake of testing, StairVisualizer is built based on
same behaviors as the ObstacleVisualizer. It contains a
cloud viewer and a PCL Visualizer which overlays stairs
on top of it.
The visualizer is still problematic. It seems on some systems
it faces "Segmentation Fault" error.
TODO: Check if downgrade to OpenCV 2.4.9 solves this problem.